Refactor bellpush code
